basemodel schema bug
