+++ aries_cloudagent/anoncreds/anoncreds/default/did_web_registry/registry.py
@@ -3,8 +3,8 @@ from ...models import (
     AnonCredsRegistryGetCredentialDefinition,
     AnonCredsRegistryGetRevocationList,
     AnonCredsRegistryGetRevocationRegistryDefinition,
-    AnonCredsRegistryGetSchema,
 )
+from ....models.anoncreds_schema import AnonCredsRegistryGetSchema
 from .....config.injection_context import InjectionContext
 from ...anoncreds_registry import BaseRegistry
